refactor y tests de reporte de nodo central
